Move `SubscriptionModel` to iap package
