+++ app/src/main/java/com/trx/consumer/screens/subscriptions/list/SubscriptionsListener.kt
@@ -1,6 +1,6 @@
 package com.trx.consumer.screens.subscriptions.list
 
-import com.trx.consumer.models.common.SubscriptionModel
+import com.trx.consumer.models.common.iap.SubscriptionModel
 
 interface SubscriptionsListener {
     fun doTapSubscription(model: SubscriptionModel)
